Fixed out_of_range exception caused in initialize_load_address while setting m_load_address

+++ src/minidbg.cpp
@@ -146,7 +146,7 @@ void debugger::initialise_load_address() {
       std::string addr;
       std::getline(map, addr, '-');
 
-      m_load_address = std::stoi(addr, 0, 16);
+      m_load_address = std::stol(addr, 0, 16);
    }
 }
